Initial support for properties
Support exporting of all public members of a go type as a property
on D-Bus. Also allows table exports to allow types other than functions
to be exported as a property as well.

If properties are are exported then the property interface is also exported
as part of the object exporting the properties.